Make SettingsViewController, add isKeyboardDismissedWhenTappedOutside
SettingsViewController provides interface to add and delete predefined answers which will be used in case the Internet connection is not available. At this stage action methods filled with temporary placeholders.

Also extend UIView with 'isKeyboardDismissedWhenTappedOutside' property. It allows adding tap gestures to the view this property is used. If this view contains the first responder (the keyboard is visible on screen) and someone taps on the view the property will stop any editing activity happening on the view.