Settings VC business logic, InMemoryStore singleton
Create SettingsViewController business logic. The Settings VC now has the ability to:
- write to some store (persistent, in-memory, etc.) predefined answers;
- get answer from store by the index;
- get number of answers stored;
- delete answer by the index.
All operations with the store VC should perform using presenter methods, and the presenter, in turn, should perform operations with the store via 'store' property. The property shouldn't expose the store itself but return the type erased 'AnyPredefinedAnswersStore' which can be retrieved using the respective method called 'eraseToAnyPredefinedAnswersStore()'

This is made basically for two reasons: to separate abstraction from implementation and to deal with some Swift limitations. You may say that we could just use the Swift 5.1 feature called 'opaque types'. But if we want a protocol type which uses an associated type to be returned with the keyword 'some' (e.g. from the property or method) the compiler can't infer the associated type from the context (it shows '(some YourProtocol).YourAssociatedType' which the compiler can't cast to 'YourImplementationOfAssociatedType' which you get typically by typealiasing when adopting the protocol) and it creates a lot of problems dealing with this.
So it was decided to use the 'type erasure' pattern here since they allow the compiler to do better type checking in compile-time and provides a necessary level of abstraction