Storable protocol, DI of the store service
Create 'PredefinedAnswersStorable' protocol and forth follow the practice of creating '...Storable' protocols in the future when new types of stores will be required. The protocol declares a single generic function called 'setPredefinedAnswersStore(_:)'. All classes and structures that need to interact with the store should adopt this protocol.
Generally, it's recommended to hide the store instance inside of the classes/structures which use it. The store instance should be set via the previously mentioned 'setPredefinedAnswersStore(_:)' method. Also, they have to declare a model type which they use when communicating with the store instance.
The store instance should be set before any of its methods will be used inside of the hosting module so it's recommended to set the store when building a module up (i.e. while execution of the 'build()' static method of the respective module builder).